Widget config etc.

+++ app/src/main/java/us/huseli/thoucylinder/viewmodels/DeleteAlbumsViewModel.kt
@@ -6,7 +6,6 @@ import kotlinx.collections.immutable.persistentListOf
 import kotlinx.collections.immutable.toImmutableList
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.distinctUntilChanged
 import kotlinx.coroutines.flow.map
 import us.huseli.retaintheme.extensions.launchOnIOThread
 import us.huseli.thoucylinder.dataclasses.album.AlbumUiState
@@ -23,7 +22,7 @@ class DeleteAlbumsViewModel @Inject constructor(
 
     val albumUiStates: StateFlow<ImmutableList<AlbumUiState>> = _albumIds.map { albumIds ->
         repos.album.listAlbumCombos(albumIds).map { it.toUiState() }.toImmutableList()
-    }.distinctUntilChanged().stateWhileSubscribed(persistentListOf())
+    }.stateWhileSubscribed(persistentListOf())
     val isImportingLocalMedia = repos.localMedia.isImportingLocalMedia
 
     fun deleteLocalAlbumFiles() {
@@ -39,6 +38,7 @@ class DeleteAlbumsViewModel @Inject constructor(
     }
 
     fun hideAlbums(
+        deleteFiles: Boolean = false,
         onGotoLibraryClick: (() -> Unit)? = null,
         onGotoAlbumClick: ((String) -> Unit)? = null,
     ) {
@@ -46,26 +46,11 @@ class DeleteAlbumsViewModel @Inject constructor(
             val firstTitle = albumUiStates.value.first().title
 
             repos.album.setAlbumsIsHidden(_albumIds.value, true)
+            if (deleteFiles) managers.library.deleteLocalAlbumFiles(_albumIds.value)
             repos.message.onHideAlbums(
                 albumCount = _albumIds.value.size,
                 firstTitle = firstTitle,
-                onUndoClick = { reAddAlbumsToLibrary(onGotoLibraryClick, onGotoAlbumClick) },
-            )
-        }
-    }
-
-    fun hideAlbumsAndDeleteFiles(
-        onGotoLibraryClick: (() -> Unit)? = null,
-        onGotoAlbumClick: ((String) -> Unit)? = null,
-    ) {
-        launchOnIOThread {
-            val firstTitle = albumUiStates.value.first().title
-
-            repos.album.setAlbumsIsHidden(_albumIds.value, true)
-            managers.library.deleteLocalAlbumFiles(_albumIds.value)
-            repos.message.onHideAlbumsAndDeleteFiles(
-                albumCount = _albumIds.value.size,
-                firstTitle = firstTitle,
+                deleteFiles = deleteFiles,
                 onUndoClick = { reAddAlbumsToLibrary(onGotoLibraryClick, onGotoAlbumClick) },
             )
         }
